requirements.txt
discord.py
bs4
praw
requests
pyfiglet
wikipedia